Generate a Users scaffold
This creates a resource called Users that has the basic CRUD actions, as
well as the backing database table and model.

This is the groundwork for the live coding exercise of changing the
single name column to be a first and last name column.

This also updates the root route to be the Users index, which is for the
sake of this exercise.